+++ SharpEngine.Core.Components/Properties/Textures/Texture.cs
@@ -1,7 +1,5 @@
 ﻿using Silk.NET.Assimp;
 using Silk.NET.OpenGL;
-using StbImageSharp;
-using System.IO;
 
 namespace SharpEngine.Core.Components.Properties.Textures;
 
@@ -13,8 +11,8 @@ public partial class Texture : IDisposable
     /// <summary>The OpenGL handle for the texture.</summary>
     public readonly uint Handle;
     public readonly TextureType Type;
-
-    private readonly string Path;
+    public readonly string Path;
+    
     private readonly GL _gl;
 
     /// <summary>
@@ -29,60 +27,5 @@ public partial class Texture : IDisposable
         Type = type;
 
         Initialize();
-    }
-
-    public void Initialize()
-    {
-        // Bind the handle
-        Use();
-
-        // OpenGL has its texture origin in the lower left corner instead of the top left corner,
-        // so we tell StbImageSharp to flip the image when loading.
-        StbImage.stbi_set_flip_vertically_on_load(1);
-
-        // Here we open a stream to the file and pass it to StbImageSharp to load.
-        using (Stream stream = System.IO.File.OpenRead(Path))
-        {
-            var image = ImageResult.FromStream(stream, ColorComponents.RedGreenBlueAlpha);
-
-            // Generate a texture
-            _gl.TexImage2D<byte>(TextureTarget.Texture2D, 0, InternalFormat.Rgba, (uint)image.Width, (uint)image.Height, 0, PixelFormat.Rgba, PixelType.UnsignedByte, image.Data.AsSpan());
-        }
-
-        // Set texture parameters
-        SetParameters();
-
-        // Generate mipmaps
-        _gl.GenerateMipmap(GLEnum.Texture2D);
-    }
-
-    public void SetParameters()
-    {
-        _gl.TexParameter(TextureTarget.Texture2D, TextureParameterName.TextureWrapS, (int)GLEnum.ClampToEdge);
-        _gl.TexParameter(TextureTarget.Texture2D, TextureParameterName.TextureWrapT, (int)GLEnum.ClampToEdge);
-        _gl.TexParameter(TextureTarget.Texture2D, TextureParameterName.TextureMinFilter, (int)GLEnum.LinearMipmapLinear);
-        _gl.TexParameter(TextureTarget.Texture2D, TextureParameterName.TextureMagFilter, (int)GLEnum.Linear);
-        _gl.TexParameter(TextureTarget.Texture2D, TextureParameterName.TextureBaseLevel, 0);
-        _gl.TexParameter(TextureTarget.Texture2D, TextureParameterName.TextureMaxLevel, 8);
-        _gl.GenerateMipmap(TextureTarget.Texture2D);
-    }
-
-    // Multiple textures can be bound, if your shader needs more than just one.
-    // If you want to do that, use GL.ActiveTexture to set which slot GL.BindTexture binds to.
-    // The OpenGL standard requires that there be at least 16, but there can be more depending on your graphics card.
-
-    /// <summary>
-    ///     Activate the texture.
-    /// </summary>
-    /// <param name="unit">The bound texture location.</param>
-    public void Use(TextureUnit unit = TextureUnit.Texture0)
-    {
-        _gl.ActiveTexture(unit);
-        _gl.BindTexture(TextureTarget.Texture2D, Handle);
-    }
-
-    public void Dispose()
-    {
-        _gl.DeleteTexture(Handle);
     }
 }
