+++ SharpEngine.Core.Components/Properties/Textures/TextureExtensions.cs
@@ -5,37 +5,17 @@ namespace SharpEngine.Core.Components.Properties.Textures;
 
 public partial class Texture
 {
-    private readonly Dictionary<string, Texture> _textureCache = [];
-
-    /// <summary>
-    ///     Loads a texture from the specified path.
-    /// </summary>
-    /// <param name="path">the full path to the texture.</param>
-    /// <returns>The loaded texture program.</returns>
-    public Texture LoadTexture(string path)
+    public void Initialize()
     {
-        if (string.IsNullOrWhiteSpace(path))
-            throw new FileNotFoundException("No texture file provided.");
-
-        if (!File.Exists(path))
-            throw new FileNotFoundException($"Texture file not found from path '{path}'.");
-
-        // Check if the texture is already in the cache
-        if (_textureCache.TryGetValue(path, out var cachedTexture))
-            return cachedTexture;
-
-        // Generate texture
-        var texture = new Texture(_gl, path, Silk.NET.Assimp.TextureType.Diffuse);
-
         // Bind the handle
-        texture.Use();
+        Use();
 
         // OpenGL has its texture origin in the lower left corner instead of the top left corner,
         // so we tell StbImageSharp to flip the image when loading.
         StbImage.stbi_set_flip_vertically_on_load(1);
 
         // Here we open a stream to the file and pass it to StbImageSharp to load.
-        using (Stream stream = File.OpenRead(path))
+        using (Stream stream = File.OpenRead(Path))
         {
             var image = ImageResult.FromStream(stream, ColorComponents.RedGreenBlueAlpha);
 
@@ -44,15 +24,40 @@ public partial class Texture
         }
 
         // Set texture parameters
-        texture.SetParameters();
+        SetParameters();
 
         // Generate mipmaps
         _gl.GenerateMipmap(GLEnum.Texture2D);
+    }
 
-        // Add it to the cache
-        _textureCache[path] = texture;
+    public void SetParameters()
+    {
+        _gl.TexParameter(TextureTarget.Texture2D, TextureParameterName.TextureWrapS, (int)GLEnum.ClampToEdge);
+        _gl.TexParameter(TextureTarget.Texture2D, TextureParameterName.TextureWrapT, (int)GLEnum.ClampToEdge);
+        _gl.TexParameter(TextureTarget.Texture2D, TextureParameterName.TextureMinFilter, (int)GLEnum.LinearMipmapLinear);
+        _gl.TexParameter(TextureTarget.Texture2D, TextureParameterName.TextureMagFilter, (int)GLEnum.Linear);
+        _gl.TexParameter(TextureTarget.Texture2D, TextureParameterName.TextureBaseLevel, 0);
+        _gl.TexParameter(TextureTarget.Texture2D, TextureParameterName.TextureMaxLevel, 8);
+        _gl.GenerateMipmap(TextureTarget.Texture2D);
+    }
 
-        return texture;
+    // Multiple textures can be bound, if your shader needs more than just one.
+    // If you want to do that, use GL.ActiveTexture to set which slot GL.BindTexture binds to.
+    // The OpenGL standard requires that there be at least 16, but there can be more depending on your graphics card.
+
+    /// <summary>
+    ///     Activate the texture.
+    /// </summary>
+    /// <param name="unit">The bound texture location.</param>
+    public void Use(TextureUnit unit = TextureUnit.Texture0)
+    {
+        _gl.ActiveTexture(unit);
+        _gl.BindTexture(TextureTarget.Texture2D, Handle);
+    }
+
+    public void Dispose()
+    {
+        _gl.DeleteTexture(Handle);
     }
 
 }
